Initial commit projektu Shop.js

+++ controllers/admin.js
@@ -1,5 +1,4 @@
 const { validationResult } = require('express-validator');
-
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
@@ -15,10 +14,27 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const errors = validationResult(req);
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description,
+        imageUrl: ''
+      },
+      errorMessage: 'Attached file is not a valid image (only .png, .jpg, .jpeg allowed).',
+      validationErrors: []
+    });
+  }
 
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
@@ -28,20 +44,22 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
-        description: description
+        description: description,
+        imageUrl: ''
       },
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array()
     });
   }
+  
+  const imageUrl = image.path;
 
   const product = new Product({
     title: title,
     price: price,
     description: description,
-    imageUrl: imageUrl,
+    imageUrl: imageUrl, 
     userId: req.user
   });
 
@@ -88,7 +106,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedPrice = req.body.price;
   const updatedDesc = req.body.description;
 
@@ -102,7 +120,7 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
+        imageUrl: image ? image.path : req.body.oldImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -120,7 +138,12 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      
+      if (image) {
+        // Logika usuwania starego pliku powinna być tutaj
+        product.imageUrl = image.path;
+      }
+      
       return product.save().then(result => {
         res.redirect('/admin/products');
       });
@@ -150,7 +173,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      // Logika usuwania pliku powinna być tutaj
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       res.redirect('/admin/products');
     })
